move carousel to app/landing page
