+++ frontend/src/modules/app/landing-page/Carousel.tsx
