+++ frontend/src/modules/controllers/ValueDisplay.tsx
