+++ frontend/src/modules/controllers/ValueSelectorDisplay.tsx
