--- frontend/src/modules/displays/index.ts
@@ -1,15 +1,0 @@
-/**
- * @summary UI resuable components used in app
- *
- * @file Components which can be reused across app
- *
- * Currently it has Carousel(used in Landing page to show slides of content)
- * ModeBanner(Current mode text shown in footer of quickstart./set alarms page)
- * ToogleValue(Used in Settings page to toggle values for theme color & unit)
- * ValueDisplay(All the numeric values shown across app uses this component )
- * ValueSelectorDisplay (Its similar to `ValueDisplay` except that it accepts
- * selector as input & shows the value fetching from redux store using selector )
- */
-export * from './Carousel';
-export * from './ValueDisplay';
-export * from './ValueSelectorDisplay';
