+++ frontend/src/store/controller/selectors.ts
@@ -45,7 +45,7 @@ import {
   PVPoint,
   ControllerStates,
 } from './types';
-import { SelectorType } from '../../modules/displays/ValueSelectorDisplay';
+import { SelectorType } from '../../modules/controllers/ValueSelectorDisplay';
 
 export const roundValue = (value: number): number => {
   return value !== undefined && !Number.isNaN(value)
